requirements.txt
aiocontextvars==0.2.2
click==7.1.2
contextvars==2.4
emoji==0.6.0
immutables==0.14
loguru==0.5.1
mysql-connector-python==8.0.21
pkg-resources==0.0.0
protobuf==3.13.0
six==1.15.0
